requirements.txt
 python-telegram-bot>=20.0
sqlalchemy>=2.0
python-dotenv>=1.0